+++ build_helpers/docker/build_arm64_static_local.sh
@@ -13,8 +13,13 @@
 #  * Try running an arm64 executable:
 #    $ docker run --rm -t arm64v8/ubuntu uname -m 
 
+# NOTE on Alpine v3.17:
+# * Alpine v3.18-3.20 fail on ARM64 with git clone error "bad address".
+# * Alpine v3.20+ requires "apk add c-ares-static zstd-static";
+#   see Makefile alpine 3.20 hint for linker flags
+
 CONTAINER_NAME="elbencho-static"
-IMAGE_NAME="alpine:3"
+IMAGE_NAME="alpine:3.17"           # see note on alpine 3.17 above
 ELBENCHO_VERSION=$(make version)
 
 ALTHTTPSVC_SUPPORT="${OVERRIDE_ALTHTTPSVC_SUPPORT:-0}"
@@ -24,20 +29,20 @@ USE_MIMALLOC="${OVERRIDE_USE_MIMALLOC:-1}"
 # Cross-compile is mem intensive, so limit to one proc per GB RAM
 RAM_GB=$(grep MemTotal /proc/meminfo | awk '{printf "%."d"f\n", $2 / 1024 / 1024}')
 RAM_GB=$((RAM_GB > 1 ? RAM_GB : 1))
-NUM_JOBS=$(( $(nproc) > RAM_GB ? RAM_GB : $(nproc) ))
+NUM_JOBS=${NUM_JOBS:="$(( $(nproc) > RAM_GB ? RAM_GB : $(nproc) ))"}
 
-echo "Calculated NUM_JOBS: $NUM_JOBS (RAM_GB: $RAM_GB, nproc: $(nproc))"
+echo "NUM_JOBS: $NUM_JOBS (RAM_GB: $RAM_GB, nproc: $(nproc))"
 
 docker rm $CONTAINER_NAME
 
 #docker pull $IMAGE_NAME && \
-docker run --platform linux/arm64 --name $CONTAINER_NAME --privileged -it -v $PWD:$PWD -w $PWD $IMAGE_NAME \
+docker run --platform linux/arm64 --name $CONTAINER_NAME --privileged -i -v $PWD:$PWD -w $PWD $IMAGE_NAME \
     sh -c "\
     apk add bash boost-dev build-base gcc g++ git libaio-dev make numactl-dev \
-        c-ares-static cmake curl-dev curl-static openssl-libs-static ncurses-static \
+        cmake curl-dev curl-static openssl-libs-static ncurses-static \
         boost-static ncurses zlib-static libretls-static nghttp2-static \
         brotli-static ncurses-dev sudo tar libidn2-static libunistring-static \
-        libpsl-static zstd-static && \
+        libpsl-static && \
     apk update && apk upgrade && \
     adduser -u $UID -D builduser && \
     sudo -u builduser git config --global submodule.fetchJobs $NUM_JOBS && \