+++ build_helpers/docker/build_static_local.sh
@@ -7,8 +7,12 @@
 # Note: For podman's docker interface (e.g. RHEL8) remove "sudo -u builduser",
 #       because it already runs under the calling UID.
 
+# NOTE on Alpine v3.17:
+# * Alpine v3.18-3.20 fail on ARM64 with git clone error "bad address".
+# * Alpine v3.20+ requires "apk add c-ares-static zstd-static";
+#   see Makefile alpine 3.20 hint for linker flags.
 CONTAINER_NAME="elbencho-static"
-IMAGE_NAME="alpine:3"
+IMAGE_NAME="alpine:3.17"            # see note on alpine 3.17 above
 ELBENCHO_VERSION=$(make version)
 
 ALTHTTPSVC_SUPPORT="${OVERRIDE_ALTHTTPSVC_SUPPORT:-1}"
@@ -20,13 +24,13 @@ rm -f packaging/elbencho-\${ELBENCHO_VERSION}-static-$(uname -m).tar.gz
 docker rm $CONTAINER_NAME
 
 docker pull $IMAGE_NAME && \
-docker run --name $CONTAINER_NAME --privileged -it -v $PWD:$PWD -w $PWD $IMAGE_NAME \
+docker run --name $CONTAINER_NAME --privileged -i -v $PWD:$PWD -w $PWD $IMAGE_NAME \
     sh -c "\
     apk add bash boost-dev build-base gcc g++ git libaio-dev make numactl-dev \
-        c-ares-static cmake curl-dev curl-static openssl-libs-static ncurses-static \
+        cmake curl-dev curl-static openssl-libs-static ncurses-static \
         boost-static ncurses zlib-static libretls-static nghttp2-static \
         brotli-static ncurses-dev sudo tar libidn2-static libunistring-static \
-        libpsl-static zstd-static && \
+        libpsl-static && \
     apk update && apk upgrade && \
     adduser -u $UID -D builduser && \
     sudo -u builduser git config --global submodule.fetchJobs 0 && \